+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 2.0.3
+- Partial operations config
+
+# 2.0.2
+- Missed media fix
+
 # 2.0.1
 - Plugin config payment methods interface update